Move modifiers to its own files
