+++ podcasts/Styles.swift
@@ -352,47 +352,3 @@ struct ModalTopPill: View {
         static let pillOpacity: CGFloat = 0.2
     }
 }
-
-// MARK: - Mini Player padding
-
-/// Apply a bottom padding whenever the mini player is visible
-struct MiniPlayerPadding: ViewModifier {
-    @State var isMiniPlayerVisible: Bool = false
-
-    func body(content: Content) -> some View {
-        content
-            .padding(.bottom, isMiniPlayerVisible ? Constants.Values.miniPlayerOffset - 2 : 0).onAppear {
-                isMiniPlayerVisible = (PlaybackManager.shared.currentEpisode() != nil)
-            }
-            .ignoresSafeArea(.keyboard)
-            .onReceive(NotificationCenter.default.publisher(for: Constants.Notifications.miniPlayerDidAppear), perform: { _ in
-                isMiniPlayerVisible = true
-            })
-            .onReceive(NotificationCenter.default.publisher(for: Constants.Notifications.miniPlayerDidDisappear), perform: { _ in
-                isMiniPlayerVisible = false
-            })
-    }
-}
-
-// MARK: - Dismiss keyboard when scrolling
-
-extension UIApplication {
-    func endEditing(_ force: Bool) {
-        (connectedScenes.first as? UIWindowScene)?.windows.filter { $0.isKeyWindow }.first?.endEditing(force)
-    }
-}
-
-struct DismissKeyboardOnScroll: ViewModifier {
-    var gesture = DragGesture().onChanged { _ in
-        UIApplication.shared.endEditing(true)
-    }
-    func body(content: Content) -> some View {
-        if #available(iOS 16.0, *) {
-            content
-                .scrollDismissesKeyboard(.interactively)
-        } else {
-            content
-                .highPriorityGesture(gesture)
-        }
-    }
-}
